Modifying Food items and adding new food items.
I have finally edited the food items and added the new food items, together with the new images and selecting the size of the food items.
And then, I tried testing the new food items and it works fine. Suddenly, I realized that there is something wrong with adding the non-drinks food items so i finally fixed it.