I added the ability to update the status of the order.
Anyways, I successfully created the whole full stack webpage of Food Ordering System. But that's not all, I will enhance or improve the webpage on the next sprint.

These are include: Tracking the order, modifying a new food item, and adding a new category. Another thing is creating several new features for the interface for the admin panel.